requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
atomicwrites==1.4.1
attrs==22.2.0
certifi==2024.12.14a
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
dataclasses==0.8
dnspython==2.2.1
email-validator==1.3.1
Faker==14.2.1
idna==3.10
importlib-metadata==4.8.3
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.2
pyparsing==3.1.4
pytest==7.0.1
python-dateutil==2.9.0.post0
questions==0.7.1
requests==2.27.1
selenium==3.141.0
simpleeval==0.9.13
six==1.17.0
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.20
zipp==3.6.0
